# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stock Simulator Service
  description: Health/ready endpoints and Prometheus metrics endpoint.
  version: 0.1.0
paths:
  /healthz:
    get:
      summary: Healthz
      operationId: healthz_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthzResponse'
  /readyz:
    get:
      summary: Readyz
      operationId: readyz_readyz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyzResponse'
  /v1/reset:
    post:
      summary: Reset
      operationId: reset_v1_reset_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/observe:
    get:
      summary: Observe
      operationId: observe_v1_observe_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObserveResponse'
  /v1/market_window:
    get:
      summary: Market Window
      operationId: market_window_v1_market_window_get
      parameters:
      - name: start
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Start
      - name: end
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketWindowContentModel'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/step_many:
    post:
      summary: Step Many
      operationId: step_many_v1_step_many_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepManyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepManyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    EncodedActionModel:
      properties:
        side_code:
          type: integer
          title: Side Code
        units:
          type: number
          title: Units
        order_type_code:
          type: integer
          title: Order Type Code
        has_limit_price:
          type: boolean
          title: Has Limit Price
        limit_price:
          anyOf:
          - type: number
          - type: 'null'
          title: Limit Price
      additionalProperties: false
      type: object
      required:
      - side_code
      - units
      - order_type_code
      - has_limit_price
      title: EncodedActionModel
      description: Transport model for encoded ``step_many`` action rows.
    EnvStateModel:
      properties:
        t:
          type: integer
          title: T
        cash:
          type: number
          title: Cash
        units:
          type: number
          title: Units
        equity:
          type: number
          title: Equity
        leverage:
          type: number
          title: Leverage
        open_orders:
          type: integer
          title: Open Orders
        done:
          type: boolean
          title: Done
      additionalProperties: false
      type: object
      required:
      - t
      - cash
      - units
      - equity
      - leverage
      - open_orders
      - done
      title: EnvStateModel
      description: Pydantic transport model for :class:`EnvState`.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HealthzResponse:
      properties:
        status:
          type: string
          const: ok
          title: Status
      additionalProperties: false
      type: object
      required:
      - status
      title: HealthzResponse
      description: Liveness response payload.
    MarketWindowContentModel:
      properties:
        start:
          type: integer
          title: Start
        end:
          type: integer
          title: End
        t:
          type: integer
          title: T
        rows:
          $ref: '#/components/schemas/MarketWindowRowsModel'
      additionalProperties: false
      type: object
      required:
      - start
      - end
      - t
      - rows
      title: MarketWindowContentModel
      description: Pydantic transport model for :class:`MarketWindowContent`.
    MarketWindowRowsModel:
      properties:
        open:
          items:
            type: number
          type: array
          title: Open
        high:
          items:
            type: number
          type: array
          title: High
        low:
          items:
            type: number
          type: array
          title: Low
        close:
          items:
            type: number
          type: array
          title: Close
        volume:
          items:
            type: number
          type: array
          title: Volume
      additionalProperties: false
      type: object
      required:
      - open
      - high
      - low
      - close
      - volume
      title: MarketWindowRowsModel
      description: Per-bar OHLCV column arrays for a market window.
    MarketWindowViewHandleModel:
      properties:
        start:
          type: integer
          title: Start
        end:
          type: integer
          title: End
        t:
          type: integer
          title: T
        current_price:
          type: number
          title: Current Price
      additionalProperties: false
      type: object
      required:
      - start
      - end
      - t
      - current_price
      title: MarketWindowViewHandleModel
      description: Pydantic transport model for :class:`MarketWindowViewHandle`.
    ObservationModel:
      properties:
        market_window_handle:
          $ref: '#/components/schemas/MarketWindowViewHandleModel'
        portfolio_vector:
          items:
            type: number
          type: array
          title: Portfolio Vector
        order_summary_vector:
          items:
            type: number
          type: array
          title: Order Summary Vector
        done:
          type: boolean
          title: Done
      additionalProperties: false
      type: object
      required:
      - market_window_handle
      - portfolio_vector
      - order_summary_vector
      - done
      title: ObservationModel
      description: Pydantic transport model for :class:`Observation`.
    ObserveResponse:
      properties:
        observation:
          $ref: '#/components/schemas/ObservationModel'
      additionalProperties: false
      type: object
      required:
      - observation
      title: ObserveResponse
      description: Observe response payload.
    ReadyzResponse:
      properties:
        status:
          type: string
          enum:
          - ready
          - not_ready
          title: Status
        engine_enabled:
          type: boolean
          title: Engine Enabled
        engine_ready:
          type: boolean
          title: Engine Ready
      additionalProperties: false
      type: object
      required:
      - status
      - engine_enabled
      - engine_ready
      title: ReadyzResponse
      description: Readiness response payload.
    ResetRequest:
      properties:
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
        start_t:
          anyOf:
          - type: integer
          - type: 'null'
          title: Start T
      additionalProperties: false
      type: object
      title: ResetRequest
      description: Reset request payload.
    ResetResponse:
      properties:
        state:
          $ref: '#/components/schemas/EnvStateModel'
      additionalProperties: false
      type: object
      required:
      - state
      title: ResetResponse
      description: Reset response payload.
    StepManyRequest:
      properties:
        actions:
          items:
            $ref: '#/components/schemas/EncodedActionModel'
          type: array
          title: Actions
        include_trace:
          type: boolean
          title: Include Trace
          default: false
      additionalProperties: false
      type: object
      required:
      - actions
      title: StepManyRequest
      description: Step-many request payload.
    StepManyResponse:
      properties:
        observations:
          items:
            $ref: '#/components/schemas/ObservationModel'
          type: array
          title: Observations
        rewards:
          items:
            type: number
          type: array
          title: Rewards
        dones:
          items:
            type: boolean
          type: array
          title: Dones
        trace:
          items:
            $ref: '#/components/schemas/StepTraceRowModel'
          type: array
          title: Trace
      additionalProperties: false
      type: object
      required:
      - observations
      - rewards
      - dones
      - trace
      title: StepManyResponse
      description: Step-many response payload.
    StepTraceRowModel:
      properties:
        index:
          type: integer
          title: Index
        side_code:
          type: integer
          title: Side Code
        requested_units:
          type: number
          title: Requested Units
        order_type_code:
          type: integer
          title: Order Type Code
        has_limit_price:
          type: boolean
          title: Has Limit Price
        limit_price:
          anyOf:
          - type: number
          - type: 'null'
          title: Limit Price
        fills:
          type: integer
          title: Fills
        reward:
          type: number
          title: Reward
        done:
          type: boolean
          title: Done
        t:
          type: integer
          title: T
        cash:
          type: number
          title: Cash
        position_units:
          type: number
          title: Position Units
        equity:
          type: number
          title: Equity
        leverage:
          type: number
          title: Leverage
        open_orders:
          type: integer
          title: Open Orders
        market_price:
          type: number
          title: Market Price
      additionalProperties: false
      type: object
      required:
      - index
      - side_code
      - requested_units
      - order_type_code
      - has_limit_price
      - fills
      - reward
      - done
      - t
      - cash
      - position_units
      - equity
      - leverage
      - open_orders
      - market_price
      title: StepTraceRowModel
      description: Pydantic transport model for :class:`StepTraceRow`.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError